Added fraction year length IO
git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism/glimmer/trunk@395 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

+++ src/fortran/glimmer_ncdf.f90
@@ -102,9 +102,9 @@ module glimmer_ncdf
 
      type(glimmer_nc_stat) :: nc
      !*FD structure containg file info
-     integer :: freq=1000
+     real :: freq=1000
      !*FD frequency at which data is written to file
-     integer :: next_write=0
+     real :: next_write=0
      !*FD next time step at which data is dumped
      integer :: timecounter=1
      !*FD time counter
